+++ src/main/java/org/betonquest/betonquest/feature/journal/JournalMainPageEntry.java
@@ -2,6 +2,7 @@ package org.betonquest.betonquest.feature.journal;
 
 import org.betonquest.betonquest.api.message.Message;
 import org.betonquest.betonquest.id.ConditionID;
+import org.betonquest.betonquest.instruction.variable.Variable;
 
 import java.util.List;
 
@@ -12,5 +13,5 @@ import java.util.List;
  * @param conditions the conditions to display the entry
  * @param entry      the message content
  */
-public record JournalMainPageEntry(int priority, List<ConditionID> conditions, Message entry) {
+public record JournalMainPageEntry(int priority, Variable<List<ConditionID>> conditions, Message entry) {
 }