+++ src/main/java/org/betonquest/betonquest/message/ParsedSectionMessage.java
@@ -8,7 +8,6 @@ import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.instruction.argument.Argument;
 import org.betonquest.betonquest.instruction.variable.Variable;
 import org.betonquest.betonquest.kernel.processor.quest.VariableProcessor;
-import org.betonquest.betonquest.variables.GlobalVariableResolver;
 import org.bukkit.configuration.ConfigurationSection;
 
 import java.util.HashMap;
@@ -46,9 +45,9 @@ public class ParsedSectionMessage extends ParsedMessage {
             return parseSection(variableProcessor, pack, section, path);
         } else if (section.isList(path)) {
             return Map.of(languageProvider.getDefaultLanguage(), new Variable<>(variableProcessor, pack,
-                    GlobalVariableResolver.resolve(pack, String.join("\n", section.getStringList(path))), Argument.STRING));
+                    String.join("\n", section.getStringList(path)), Argument.STRING));
         } else if (section.isString(path)) {
-            final String raw = GlobalVariableResolver.resolve(pack, section.getString(path));
+            final String raw = section.getString(path);
             if (raw == null) {
                 throw new QuestException("No string value for '" + path + "'!");
             }
@@ -68,10 +67,10 @@ public class ParsedSectionMessage extends ParsedMessage {
         for (final String key : subSection.getKeys(false)) {
             if (subSection.isList(key)) {
                 messages.put(key, new Variable<>(variableProcessor, pack,
-                        GlobalVariableResolver.resolve(pack, String.join("\n", subSection.getStringList(key))), Argument.STRING));
+                        String.join("\n", subSection.getStringList(key)), Argument.STRING));
                 continue;
             }
-            final String raw = GlobalVariableResolver.resolve(pack, subSection.getString(key));
+            final String raw = subSection.getString(key);
             if (raw == null) {
                 throw new QuestException("No string value for key '" + key + "'!");
             }